requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
Django==4.2
django-bootstrap5==23.4
django-widget-tweaks==1.5.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
PySocks==1.7.1
python-decouple==3.8
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
